* #10: Updated Email templates to use new identification fields and changed email file structure.     * src/emails/ConfirmEmail.ts deleted         - Previously all emails were defined here, encapsulated inside the same ConfirmEmail class.         - For better organization and maintainability, I restructured the email templates into separate         files based on their context (admin, auth, status).     * src/emails/admin/index.ts:         - Creates and common exports the admin related emails as AdminEmails class.         - Defines .send method for each email template to handle sending the email with the appropriate data.     * src/emails/auth/index.ts:         - Similar to admin/index.ts, this file defines and exports the AuthEmails class for authentication             related email templates.     * src/emails/status/PendingOrder.ts:         - Updated the email template to use the new businessId field     * Moved emails from ConfirmEmail class to their respective files in the new structure.     * Updated email templates to use the new identification fields (businessId, userId) instead of the old rut and businessRut fields.     * src/types/index.ts: Moved types for email data to ensure type safety when sending emails.
