+++ src/emails/admin/index.ts
@@ -1,0 +1,8 @@
+import { ConfirmUserEmail } from "./ConfirmUserEmail";
+
+//* Common exports
+export class AdminEmails {
+    static ConfirmUser = {
+        send: (ConfirmUserEmail.sendConfirmUserEmail)
+    }
+}
