+++ src/types/index.ts
@@ -1,0 +1,19 @@
+export interface UserEmailInterface {
+    email: string
+    name: string
+    token: string
+}
+
+export interface AdminEmailInterface {
+    userId: string;
+    email: string;
+    name: string;
+    businessName: string;
+    personalId: string;
+    businessId: string;
+    idType: string;
+    country: string;
+    address: string;
+    phone: string;
+    token: string;
+}
